v2.0.2.1

+++ src/AUTD3Sharp.cs
@@ -4,7 +4,7 @@
  * Created Date: 23/05/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 24/05/2022
+ * Last Modified: 25/05/2022
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022 Hapis Lab. All rights reserved.
@@ -425,7 +425,7 @@ namespace AUTD3Sharp
         {
             if (amp.Length != phase.Length) throw new ArgumentException();
             var length = amp.Length;
-            Base.AUTDGainCustom(out handle, amp, phase, length);
+            Base.AUTDGainCustom(out handle, amp, phase, (ulong)length);
         }
     }
